+++ examples/trainer.py
@@ -10,7 +10,7 @@ import matplotlib.pyplot as plt
 from pathlib import Path
 from torch.utils.data import DataLoader
 from torchfitter.trainer import Trainer
-from torchfitter.utils import DataWrapper
+from torchfitter.utils.data import DataWrapper
 from torchfitter.conventions import ParamsDict
 from sklearn.model_selection import train_test_split
 from torchfitter.regularization import L1Regularization